requirements.txt
Django==6.0.1
djangorestframework==3.16.1
pillow==12.1.0
psycopg2==2.9.11
python-dotenv==1.2.1
django-filter==25.2